refactor(core): split large schema files into modular structure (WIP)
- Split krd.ts (244 lines) into krd/ subdirectory
- Split target-values.ts (212 lines) into target-values/ subdirectory
- Update imports in FIT adapters
- Fix circular dependency in re-exports

Status: WIP - Need to update remaining TCX/Zwift adapter imports
