+++ packages/core/src/adapters/fit/krd-to-fit/krd-to-fit-target-power.mapper.ts
@@ -1,7 +1,5 @@
-import {
-  targetTypeSchema,
-  targetUnitSchema,
-} from "../../../domain/schemas/target";
+import { targetTypeSchema } from "../../../domain/schemas/target";
+import { targetUnitSchema } from "../../../domain/schemas/target-values";
 import type { WorkoutStep } from "../../../domain/schemas/workout";
 import { fitTargetTypeSchema } from "../schemas/fit-target";
