+++ packages/core/src/adapters/zwift/interval/interval-type-detector.ts
@@ -1,8 +1,6 @@
 import { intensitySchema } from "../../../domain/schemas/intensity";
-import {
-  targetTypeSchema,
-  targetUnitSchema,
-} from "../../../domain/schemas/target";
+import { targetTypeSchema } from "../../../domain/schemas/target";
+import { targetUnitSchema } from "../../../domain/schemas/target-values";
 import type { WorkoutStep } from "../../../domain/schemas/workout";
 
 export type ZwiftIntervalType =
