+++ packages/core/src/domain/schemas/target-values/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Target Values - Re-exports from modular structure
+ */
+
+export { cadenceValueSchema, type CadenceValue } from "./cadence";
+export { heartRateValueSchema, type HeartRateValue } from "./heart-rate";
+export { paceValueSchema, type PaceValue } from "./pace";
+export { powerValueSchema, type PowerValue } from "./power";
+export { strokeTypeValueSchema, type StrokeTypeValue } from "./stroke-type";
+export { targetUnitSchema, type TargetUnit } from "./unit";
